requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.23
pydantic[email]>=1.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic-settings>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
requests>=2.26.0
alembic>=1.7.1
bcrypt>=4.0.1
mysql-connector-python